Forked `Websocket4Net` and migrated to `Net6`
This commit adds the source code for the `Websocket4Net` nuget directly to the project.
Maybe this helps to identify the `NRE` exception.
